requirements.txt
dateutils
Django>1.8,<1.9
fabric
pep8
pyflakes
webtest
django-widget-tweaks
django-filter
django_bootstrap3
django-reversion==1.8.7
django_tables2
django_rosetta
django-grappelli
django_select2<5
django_date_extensions
django_report_builder
Pillow
django-autofixture
pytz
mysql-python
raven